+++ packages/perseus/src/widgets/explanation/explanation.tsx
@@ -10,12 +10,12 @@ import _ from "underscore";
 import {PerseusI18nContext} from "../../components/i18n-context";
 import * as Changeable from "../../mixins/changeable";
 import Renderer from "../../renderer";
-import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/explanation/prompt-utils";
+import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/explanation/explanation-ai-utils";
 import scoreNoop from "../__shared__/score-noop";
 
 import type {PerseusExplanationWidgetOptions} from "../../perseus-types";
 import type {Widget, WidgetExports, WidgetProps} from "../../types";
-import type {ExplanationPromptJSON} from "../../widget-ai-utils/explanation/prompt-utils";
+import type {ExplanationPromptJSON} from "../../widget-ai-utils/explanation/explanation-ai-utils";
 
 type RenderProps = PerseusExplanationWidgetOptions; // transform = _.identity